Implemented session handling.

+++ src/main/java/com/capgemini/controller/RestaurantController.java
@@ -2,6 +2,8 @@ package com.capgemini.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.capgemini.entities.Item;
 import com.capgemini.entities.Restaurant;
 import com.capgemini.service.IItemService;
+import com.capgemini.service.ILoginService;
 import com.capgemini.service.IRestaurantService;
 
 @RestController
@@ -26,15 +29,46 @@ public class RestaurantController {
 	@Autowired
 	IItemService itemService;
 
+	@Autowired
+	ILoginService loginService;
+
+	@GetMapping("/restaurant")
+	public ResponseEntity<Restaurant> viewRestaurant(@RequestBody Restaurant restaurant, HttpServletRequest request) {
+		// session checking
+		boolean validLogin = loginService.checkSession(request);
+		if (!validLogin) {
+			throw new IllegalArgumentException("Not logged in");
+		}
+		// todo owner check
+
+		Restaurant newRestaurant = restaurantService.viewRestaurant(restaurant);
+		if (newRestaurant == null) {
+			return new ResponseEntity("Sorry! Restaurant not available with give ID!", HttpStatus.FORBIDDEN);
+		}
+		return new ResponseEntity<>(newRestaurant, HttpStatus.OK);
+	}
+
 	@PostMapping("/restaurant")
-	public ResponseEntity<Restaurant> addRestaurant(@RequestBody Restaurant restaurant) {
+	public ResponseEntity<Restaurant> addRestaurant(@RequestBody Restaurant restaurant, HttpServletRequest request) {
+		// session checking
+		boolean validLogin = loginService.checkSession(request);
+		if (!validLogin) {
+			throw new IllegalArgumentException("Not logged in");
+		}
+
 		Restaurant saved = null;
 		saved = restaurantService.addRestaurant(restaurant);
 		return new ResponseEntity<>(saved, HttpStatus.OK);
 	}
 
 	@PutMapping("/restaurant")
-	public ResponseEntity<Restaurant> updateRestaurant(@RequestBody Restaurant restaurant) {
+	public ResponseEntity<Restaurant> updateRestaurant(@RequestBody Restaurant restaurant, HttpServletRequest request) {
+		// session checking
+		boolean validLogin = loginService.checkSession(request);
+		if (!validLogin) {
+			throw new IllegalArgumentException("Not logged in");
+		}
+
 		Restaurant newRestaurant = restaurantService.updateRestaurant(restaurant);
 		if (newRestaurant == null) {
 			return new ResponseEntity("Sorry! Restaurant not available to Update!", HttpStatus.NOT_FOUND);
@@ -43,21 +77,18 @@ public class RestaurantController {
 	}
 
 	@DeleteMapping("/restaurant")
-	public ResponseEntity<Restaurant> removeRestaurant(@RequestBody Restaurant restaurant) {
+	public ResponseEntity<Restaurant> removeRestaurant(@RequestBody Restaurant restaurant, HttpServletRequest request) {
+		// session checking
+		boolean validLogin = loginService.checkSession(request);
+		if (!validLogin) {
+			throw new IllegalArgumentException("Not logged in");
+		}
+
 		Restaurant delRestaurant = restaurantService.removeRestaurant(restaurant);
 		if (delRestaurant == null) {
 			return new ResponseEntity("Sorry! Restaurant not available!", HttpStatus.NOT_FOUND);
 		}
 		return new ResponseEntity<>(delRestaurant, HttpStatus.OK);
-	}
-
-	@GetMapping("/restaurant")
-	public ResponseEntity<Restaurant> viewRestaurant(@RequestBody Restaurant restaurant) {
-		Restaurant newRestaurant = restaurantService.viewRestaurant(restaurant);
-		if (newRestaurant == null) {
-			return new ResponseEntity("Sorry! Restaurant not available to Update!", HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<>(newRestaurant, HttpStatus.OK);
 	}
 
 	@GetMapping("/restaurant/{location}")
@@ -80,7 +111,7 @@ public class RestaurantController {
 
 	@PutMapping("/restaurant/{restaurantId}/item/{itemId}")
 	public ResponseEntity<Restaurant> addItemToRestaurant(@PathVariable String restaurantId,
-			@PathVariable String itemId) {
+			@PathVariable String itemId, HttpServletRequest request) {
 		Restaurant restaurantFromRepo = restaurantService.viewRestaurant(restaurantId);
 		Item item = itemService.viewItem(itemId);
 		restaurantFromRepo.addItem(item);
@@ -88,16 +119,4 @@ public class RestaurantController {
 		return new ResponseEntity<>(restaurantFromRepo, HttpStatus.OK);
 
 	}
-
-//	@PostMapping("/item")
-//	public ResponseEntity<Item> addItem(@RequestBody Item item) {
-//		Item returnedItem = itemService.addItem(item);
-//		return new ResponseEntity<>(returnedItem, HttpStatus.OK);
-//	}
-//
-//	@GetMapping("/item/{itemId}")
-//	public Item getAllItems(@PathVariable String itemId) {
-//		return itemService.viewItem(itemId);
-//	}
-
 }
